delete favoriteController.java, add FavoriteController.java
