+++ src/main/webapp/WEB-INF/js/userMyPage.js
@@ -1,6 +1,6 @@
 // MonoRecipe/src/main/webapp/WEB-INF/js/userMyPage.js
 
-$(document).ready(function () {
+$(function() {
     // 페이지 로드 시 즐겨찾기 메뉴의 배경색 설정
     $('#userFavoriteMenu a').css('background-color', '#222').css('color', 'white');
 
@@ -15,6 +15,15 @@ $(document).ready(function () {
         $('#userFavoriteMenu a').css('background-color', '#222').css('color', 'white'); // 즐겨찾기 메뉴 배경색 변경
         $('#userUpdateMenu a').css('background-color', 'transparent').css('color', 'black'); // 회원정보 메뉴 원래대로
     });
+    
+    
+    
+    
+    
+    
+    
+    
+    
 });
